requirements.txt
streamlit>=1.37,<2
pandas>=2.2
openpyxl>=3.1
jinja2>=3.1
python-dotenv>=1.0
google-generativeai>=0.8.3
pyarrow>=16.1
playwright>=1.55.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pyyaml>=6.0.0

# 설치 후 한번은 아래를 실행해야 브라우저가 받아집니다.
#   playwright install chromium
